src/tests/automated_tests/requirements.txt
pytest
python-dotenv
pytest-dotenv
pytest-cases
pyjwt
requests
